+++ index.d.ts
@@ -5,7 +5,8 @@ import {
 	DeleteResult as TCPDeleteResult,
 	EventAppearedCallback,
 	LiveProcessingStartedCallback,
-	SubscriptionDroppedCallback
+	SubscriptionDroppedCallback,
+	StreamEventsSlice
 }  from 'node-eventstore-client';
 
 import { 
@@ -78,6 +79,35 @@ export interface TCPWriteEventsOptions extends TCPWriteEventOptions {
 	transactionWriteSize?: number;
 }
 
+export interface TCPReadResult extends StreamEventsSlice {
+	events: Event[]
+}
+
+export interface HTTPReadResultAuthor {
+	name: string;
+}
+
+export interface HTTPReadResultLink {
+	uri: string;
+	relation: string;
+}
+
+export interface HTTPReadResult {
+	title: string,
+	id: string,
+	updated: string,
+	streamId: string,
+	author: HTTPReadResultAuthor,
+	headOfStream: boolean,
+	isEndOfStream: boolean,
+	readDirection: ReadDirection,
+	fromEventNumber: number;
+	nextEventNumber: number;
+	selfUrl: string,
+	links: HTTPReadResultLink[],
+	events: Event[]
+}
+
 export interface ProjectionStateOptions {
 	partition?: string;
 }
@@ -125,8 +155,10 @@ export class HTTPClient {
 	writeEvent(streamName: string, eventType: string, data: object, metaData?: object, options?: HTTPWriteEventOptions): Promise<void>;
 	writeEvents(streamName: string, events: NewEvent[], options?: HTTPWriteEventOptions): Promise<void>;
 	getAllStreamEvents(streamName: string, chunkSize?: number, startPosition?: number, resolveLinkTos?: boolean, embed?: EmbedType): Promise<Event[]>;
-	getEvents(streamName: string, startPosition?: number, length?: number, direction?: ReadDirection, resolveLinkTos?: boolean, embed?: EmbedType): Promise<Event[]>;
-	getEventsByType(streamName: string, eventTypes: string[], startPosition?: number, length?: number, direction?: ReadDirection, resolveLinkTos?: boolean): Promise<Event[]>;
+	getEvents(streamName: string, startPosition?: number, count?: number, direction?: ReadDirection, resolveLinkTos?: boolean, embed?: EmbedType): Promise<Event[]>;
+	getEventsByType(streamName: string, eventTypes: string[], startPosition?: number, count?: number, direction?: ReadDirection, resolveLinkTos?: boolean): Promise<Event[]>;
+	readEventsForward(streamName: string, startPosition?: number, count?: number, resolveLinkTos?: boolean, embed?: EmbedType): Promise<HTTPReadResult>;
+	readEventsBackward(streamName: string, startPosition?: number, count?: number, resolveLinkTos?: boolean, embed?: EmbedType): Promise<HTTPReadResult>;
 	deleteStream(streamName: string, hardDelete?: boolean): Promise<void>;
 	ping(): Promise<void>;
 	admin: {
@@ -161,14 +193,16 @@ export class TCPClient {
 	writeEvent(streamName: string, eventType: string, data: object, metaData?: object, options?: TCPWriteEventOptions): Promise<TCPWriteResult>;
 	writeEvents(streamName: string, events: NewEvent[], options?: TCPWriteEventsOptions): Promise<TCPWriteResult>;
 	getAllStreamEvents(streamName: string, chunkSize?: number, startPosition?: number, resolveLinkTos?: boolean): Promise<Event[]>;
-	getEvents(streamName: string, startPosition?: number, length?: number, direction?: ReadDirection, resolveLinkTos?: boolean): Promise<Event[]>;
-	getEventsByType(streamName: string, eventTypes: string[], startPosition?: number, length?: number, direction?: ReadDirection, resolveLinkTos?: boolean): Promise<Event[]>;
+	getEvents(streamName: string, startPosition?: number, count?: number, direction?: ReadDirection, resolveLinkTos?: boolean): Promise<Event[]>;
+	getEventsByType(streamName: string, eventTypes: string[], startPosition?: number, count?: number, direction?: ReadDirection, resolveLinkTos?: boolean): Promise<Event[]>;
+	readEventsForward(streamName: string, startPosition?: number, count?: number, resolveLinkTos?: boolean): Promise<TCPReadResult>;
+	readEventsBackward(streamName: string, startPosition?: number, count?: number, resolveLinkTos?: boolean): Promise<TCPReadResult>;
 	deleteStream(streamName: string, hardDelete?: boolean): Promise<TCPDeleteResult>;
 	eventEnumerator(streamName: string, direction?: ReadDirection, resolveLinkTos?: boolean): {
-		first(length: number): Promise<EventEnumeratorResult>;
-		last(length: number): Promise<EventEnumeratorResult>;
-		previous(length: number): Promise<EventEnumeratorResult>;
-		next(length: number): Promise<EventEnumeratorResult>;
+		first(count: number): Promise<EventEnumeratorResult>;
+		last(count: number): Promise<EventEnumeratorResult>;
+		previous(count: number): Promise<EventEnumeratorResult>;
+		next(count: number): Promise<EventEnumeratorResult>;
 	};
 	subscribeToStream(streamName: string, onEventAppeared?: EventAppearedCallback<EventStoreSubscription>, onDropped?: SubscriptionDroppedCallback<EventStoreSubscription>, resolveLinkTos?: boolean): Promise<EventStoreSubscription>;
 	subscribeToStreamFrom(streamName: string, fromEventNumber?: number, onEventAppeared?: EventAppearedCallback<EventStoreCatchUpSubscription>, onLiveProcessingStarted?: LiveProcessingStartedCallback, onDropped?: SubscriptionDroppedCallback<EventStoreCatchUpSubscription>, settings?: SubscribeToStreamFromSettings): Promise<EventStoreCatchUpSubscription>;