+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 3.1.0 (2019-02-07)
+
+#### Features
+
+- Add readEventsForward and readEventsBackward returning read metadata + events
+
+#### Misc
+
+- Rename "length" parameter to "count"
+
 # 3.0.3 (2019-02-06)
 
 #### TCP Client
@@ -382,13 +392,13 @@
 #### HTTP Client
 
 - Fix: writeEvents return successful if empty array given
-- Fix: any get events function will default to 4096 length if greater is requested (warning also displayed)
+- Fix: any get events function will default to 4096 count if greater is requested (warning also displayed)
 - Feature: add getAllStreamEvents function
 
 #### TCP Client
 
 - Feature: added start event number on getAllStreamEvents
-- Fix: any get events function will default to 4096 length if greater is requested (warning also displayed)
+- Fix: any get events function will default to 4096 count if greater is requested (warning also displayed)
 - Change: default chunkSize of reads from 250 to 1000 
 
 #### Tests